pyproject.toml
[project]
name = "shuri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gql[aiohttp]>=3.5.0",
    "plotly>=6.0.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.5",
    "typer>=0.15.1",
]
